requirements.txt
asgiref==3.5.2
Django==4.1.2
psycopg2-binary==2.9.4
sqlparse==0.4.3

django-model-utils==4.2.0
djangorestframework==3.14.0